+++ lib/src/ui/widget/nonvisual/nil.widget.dart
@@ -5,3 +5,4 @@ import 'package:flutter/widgets.dart';
 // 这里使用SizedBox.shrink代替了
 const NIL = SizedBox.shrink();
 const SLIVER_NIL = SliverToBoxAdapter(child: SizedBox.shrink());
+const PREFERRED_NIL = PreferredSize(preferredSize: Size.zero, child: NIL);