+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [🚀] `DateTimeX`增加`asDateRange`扩展;
 - [🚀] `DateTimeRangeX`的`contains`扩展方法, 增加`includeStart`和`includeEnd`;
 - [🚀] 增加`PREFERRED_NIL`;
+- [🚀] 增加`NumberRange`类, 处理数字范围场景;
 
 ## 0.51.4
 - [🚀] `DecoratedStack`补充`transformAlignment`参数;