+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [🚀] `DateTimeRangeX`的`contains`扩展方法, 增加`includeStart`和`includeEnd`;
 - [🚀] 增加`PREFERRED_NIL`;
 - [🚀] 增加`NumberRange`类, 处理数字范围场景;
+- [🚀] 增加`computeTextHeight`计算文本高度;
 
 ## 0.51.4
 - [🚀] `DecoratedStack`补充`transformAlignment`参数;